Update bootwin alias for dynamic grub-reboot command

+++ setup.sh
@@ -446,8 +446,20 @@ add_aliases() {
 
 cat >> "$target_rc" <<'EOF'
 
+#
+# ~/.bashrc
+#
+
+# If not running interactively, don't do anything
+[[ $- != *i* ]] && return
+fastfetch
+
+alias ls='ls --color=auto'
+alias grep='grep --color=auto'
+PS1='[\u@\h \W]\$ '
+
 # Windows shortcuts
-alias bootwin="sudo grub-reboot 'Windows Boot Manager (på /dev/nvme0n1p3)' && reboot"
+alias bootwin="sudo sh -c 'grub-reboot \"\$(sudo grep -i \"menuentry.*Windows Boot Manager\" /boot/grub/grub.cfg | cut -d\"'\''\" -f2 | head -n1)\" && reboot'"
 alias cleaninstall="sudo grub-reboot arch-installer && reboot"
 alias mountwin="sudo mkdir -p /run/media/methu/Windows && sudo ntfs-3g /dev/nvme0n1p2 /run/media/methu/Windows"
 alias enable_warp-svc="sudo systemctl unmask warp-svc && sudo systemctl start warp-svc"
@@ -456,6 +468,7 @@ alias disablecores='for cpu in {4..7}; do echo 0 | sudo tee /sys/devices/system/
 alias enablecores='for cpu in {4..7}; do echo 1 | sudo tee /sys/devices/system/cpu/cpu$cpu/online >/dev/null; done'
 alias cpu-cool='sudo cpupower frequency-set -u 800MHz'
 alias cpu-normal='sudo cpupower frequency-set -u 1600MHz'
+eval "$(zoxide init bash)"
 
 EOF
